darkSky API call functional, mapping data to objects, and storing in Realm successfully. Beginning to attemp tieing data to UI beginning with details screen

+++ app/src/main/java/com/andrewm/weatherornot/data/model/DailyData.kt
@@ -32,7 +32,7 @@ open class DailyData : RealmObject() {
     open var cloudCover: Double? = 0.0
     open var uvIndex: Int? = 0
     open var uvIndexTime: Int? = 0
-    open var visibility: Int? = 0
+    open var visibility: Double? = 0.0
     open var ozone: Double? = 0.0
     open var temperatureMin: Double? = 0.0
     open var temperatureMinTime: Int? = 0
